+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - `impl Listen for Cell`
 - `impl Listen for CellWithLocal`
 
+### Changed
+
+- Renamed `Sink` to `Log`, to clarify that it is not the opposite of `Source`.
+
 ## 0.1.0 (2024-12-19)
 
 Initial public release.